+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+- **The stop button of a radio station is red.** The power icon in the Player header was drawn in the header's text colour like the other controls; it is now red in either theme, whatever the header's tint (user: “the button of shutdown in radio have to be in red”).
+
 ## [4.8.1] - 2026-09-22
 
 ### Added